refactor: move add local route to seperate function
